+++ cmd/operator/main.go
@@ -25,6 +25,7 @@ import (
 	"os"
 	"os/signal"
 	"regexp"
+	"strings"
 	"syscall"
 
 	"github.com/blang/semver/v4"
@@ -36,6 +37,7 @@ import (
 	"golang.org/x/sync/errgroup"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
+	discoveryv1 "k8s.io/api/discovery/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/kubernetes"
@@ -118,9 +120,11 @@ var (
 	serverConfig = server.DefaultConfig(":8080", false)
 
 	// Parameters for the kubelet endpoints controller.
-	kubeletObject       string
-	kubeletSelector     operator.LabelSelector
-	nodeAddressPriority operator.NodeAddressPriority
+	kubeletObject        string
+	kubeletSelector      operator.LabelSelector
+	nodeAddressPriority  operator.NodeAddressPriority
+	kubeletEndpoints     bool
+	kubeletEndpointSlice bool
 
 	featureGates = k8sflag.NewMapStringBool(ptr.To(map[string]bool{}))
 )
@@ -140,6 +144,8 @@ func parseFlags(fs *flag.FlagSet) {
 	fs.StringVar(&kubeletObject, "kubelet-service", "", "Service/Endpoints object to write kubelets into in format \"namespace/name\"")
 	fs.Var(&kubeletSelector, "kubelet-selector", "Label selector to filter nodes.")
 	fs.Var(&nodeAddressPriority, "kubelet-node-address-priority", "Node address priority used by kubelet. Either 'internal' or 'external'. Default: 'internal'.")
+	fs.BoolVar(&kubeletEndpointSlice, "kubelet-endpointslice", false, "Create EndpointSlice objects for kubelet targets.")
+	fs.BoolVar(&kubeletEndpoints, "kubelet-endpoints", true, "Create Endpoints objects for kubelet targets.")
 
 	// The Prometheus config reloader image is released along with the
 	// Prometheus Operator image, tagged with the same semver version. Default to
@@ -519,15 +525,55 @@ func run(fs *flag.FlagSet) int {
 
 	var kec *kubelet.Controller
 	if kubeletObject != "" {
+		opts := []kubelet.ControllerOption{kubelet.WithNodeAddressPriority(nodeAddressPriority.String())}
+
+		kubeletService := strings.Split(kubeletObject, "/")
+		if len(kubeletService) != 2 {
+			logger.Error(fmt.Sprintf("malformatted kubelet object string %q, must be in format \"namespace/name\"", kubeletObject))
+			cancel()
+			return 1
+		}
+
+		if kubeletEndpointSlice {
+			allowed, errs, err := k8sutil.IsAllowed(
+				ctx,
+				kclient.AuthorizationV1().SelfSubjectAccessReviews(),
+				[]string{kubeletService[0]},
+				k8sutil.ResourceAttribute{
+					Group:    discoveryv1.SchemeGroupVersion.Group,
+					Version:  discoveryv1.SchemeGroupVersion.Version,
+					Resource: "endpointslices",
+					Verbs:    []string{"get", "list", "create", "update", "delete"},
+				})
+			if err != nil {
+				logger.Error(fmt.Sprintf("failed to check permissions on resource 'endpointslices' (group %q)", discoveryv1.SchemeGroupVersion.Group), "err", err)
+				cancel()
+				return 1
+			}
+
+			if !allowed {
+				for _, reason := range errs {
+					logger.Warn(fmt.Sprintf("missing permission on resource 'endpointslices' (group: %q)", discoveryv1.SchemeGroupVersion.Group), "reason", reason)
+				}
+			} else {
+				opts = append(opts, kubelet.WithEndpointSlice())
+			}
+		}
+
+		if kubeletEndpoints {
+			opts = append(opts, kubelet.WithEndpoints())
+		}
+
 		if kec, err = kubelet.New(
 			logger.With("component", "kubelet_endpoints"),
 			kclient,
 			r,
-			kubeletObject,
+			kubeletService[1],
+			kubeletService[0],
 			kubeletSelector,
 			cfg.Annotations,
 			cfg.Labels,
-			nodeAddressPriority,
+			opts...,
 		); err != nil {
 			logger.Error("instantiating kubelet endpoints controller failed", "err", err)
 			cancel()